requirements.txt
# 依赖清单

# 核心依赖
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-ollama
langgraph

# 向量数据库
pymilvus

# BM25 关键词检索
rank-bm25
jieba

# 文档解析
pdfplumber
python-docx

# 其他
python-dotenv
requests
